sql_app/requirements.txt
fastapi
fastapi_jwt_auth
sqlalchemy
PyMySQL
pydantic[email]
uvicorn
alembic
passlib
bcrypt
fastapi_mail
starlette
uuid
python-dotenv
python-multipart
python-jose[cryptography]
cloudinary==1.24.0
sqlalchemy_utils